requirements.txt
CairoSVG==2.5.2
celery==5.2.7
click==8.1.3
Flask==2.1.2
Flask-Cache==0.13.1
Flask-Login==0.6.1
Flask-Mail==0.9.1
Flask-OpenID==1.3.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
mysqlclient==2.1.0
Pillow==9.1.1
redis==4.3.3
requests==2.28.0